+++ src/sidekick/module.js
@@ -960,18 +960,46 @@
         this.status.apiUrl = apiUrl.toString();
       }
       fetch(this.status.apiUrl, { cache: 'no-store' })
-        .then((resp) => resp.json())
+        .then((resp) => {
+          // check for error status
+          if (!resp.ok) {
+            let msg = '';
+            switch (resp.status) {
+              case 401:
+                msg = 'You are not authorized to access this page. Check your login or network status.';
+                break;
+              case 404:
+                msg = this.isEditor()
+                  ? 'Page not found. Check your Sidekick configuration and make sure Helix has access to this document.'
+                  : 'Page not found. Check your Sidekick configuration or URL.';
+                break;
+              default:
+                msg = 'Failed to fetch the page status. Please try again later.';
+            }
+            throw new Error(`${resp.status}: ${msg}`);
+          }
+          return resp;
+        })
+        .then(async (resp) => {
+          try {
+            return resp.json();
+          } catch (e) {
+            throw new Error('Invalid server response. Check your Sidekick configuration or URL.');
+          }
+        })
         .then((json) => Object.assign(this.status, json))
         .then((json) => fireEvent(this, 'statusfetched', json))
         .catch((e) => {
           this.status.error = e.message;
-          this.showModal('Failed to fetch status. Please try again later', false, 0, () => {
+          this.showModal(e.message, true, 0, () => {
             // this error is fatal, hide and delete sidekick
-            window.hlx.sidekick.hide();
-            window.hlx.sidekick.replaceWith(''); // remove() doesn't work for custom element
-            delete window.hlx.sidekick;
+            if (window.hlx.sidekick) {
+              window.hlx.sidekick.hide();
+              window.hlx.sidekick.replaceWith(''); // remove() doesn't work for custom element
+              delete window.hlx.sidekick;
+            }
           });
-          console.error('failed to fetch status', e);
+          console.error('failed to fetch status', e.message);
         });
       return this;
     }
@@ -1030,7 +1058,9 @@
       if (!this.root.classList.contains('hlx-sk-hidden')) {
         this.root.classList.add('hlx-sk-hidden');
       }
-      this.hideModal();
+      if (this._modal && !this._modal.parentNode.classList.contains('hlx-sk-hidden')) {
+        this.hideModal();
+      }
       if (this.config.pushDown
         && this.hasAttribute('pushdown')
         && this.location.host !== 'docs.google.com') {
@@ -1258,6 +1288,7 @@
      */
     // eslint-disable-next-line default-param-last
     showModal(msg, sticky = false, level = 2, callback) {
+      this._modalCallback = callback;
       if (!this._modal) {
         const $spinnerWrap = appendTag(this.shadowRoot, {
           tag: 'div',
@@ -1288,9 +1319,6 @@
         const sk = this;
         window.setTimeout(() => {
           sk.hideModal();
-          if (callback && typeof callback === 'function') {
-            callback(sk);
-          }
         }, 3000);
       } else {
         this._modal.classList.add('wait');
@@ -1310,6 +1338,10 @@
         this._modal.className = '';
         this._modal.parentNode.classList.add('hlx-sk-hidden');
         fireEvent(this, 'modalhidden');
+      }
+      if (typeof this._modalCallback === 'function') {
+        this._modalCallback(this);
+        delete this._modalCallback;
       }
       return this;
     }